Renamed module containing MonadicParser
